feat: set View as abstract class and update deepCreateInstanceArrayOrPlain static method to avoid circular dependencies
