+++ ui/components/core/src/lib/types/dataset/views/Image.ts
@@ -11,7 +11,15 @@ const imageSchema = z
   })
   .passthrough();
 
+const sequenceFrameSchema = z
+  .object({
+    timestamp: z.number(),
+    frame_index: z.number(),
+  })
+  .passthrough();
+
 export type ImageType = z.infer<typeof imageSchema>; //export if needed
+export type SequenceFrameType = z.infer<typeof sequenceFrameSchema>; //export if needed
 
 export class Image extends View {
   declare data: ImageType & ViewType;
@@ -28,5 +36,20 @@ export class Image extends View {
   static nonFeaturesFields(): string[] {
     //return super.nonFeaturesFields().concat(["url", "width", "height", "format"]);
     return super.nonFeaturesFields().concat(["url"]);
+  }
+}
+
+export class SequenceFrame extends Image {
+  declare data: SequenceFrameType & ImageType & ViewType;
+
+  constructor(obj: BaseDataFields<SequenceFrameType>) {
+    if (obj.table_info.base_schema !== "SequenceFrame") throw new Error("Not a SequenceFrame");
+    sequenceFrameSchema.parse(obj.data);
+    super(obj as unknown as BaseDataFields<ImageType>);
+    this.data = obj.data as SequenceFrameType & ImageType & ViewType;
+  }
+
+  static nonFeaturesFields(): string[] {
+    return super.nonFeaturesFields().concat(["timestamp", "frame_index"]);
   }
 }
