+++ ui/components/core/src/lib/types/dataset/views/index.ts
@@ -1,4 +1,3 @@
 export * from "./Image";
-export * from "./SequenceFrame";
 export * from "./View";
 export * from "./ViewEmbedding";
