+++ ui/components/core/src/lib/utils/views/createTypedView.ts
@@ -1,0 +1,20 @@
+import {
+  Image,
+  SequenceFrame,
+  type BaseDataFields,
+  type ImageType,
+  type SequenceFrameType,
+  type ViewType,
+} from "../../types";
+
+export const createTypedView = (view: BaseDataFields<ViewType>) => {
+  if (Array.isArray(view)) {
+    const sequenceFrames: SequenceFrame[] = [];
+    for (const v of view) {
+      sequenceFrames.push(new SequenceFrame(v as unknown as BaseDataFields<SequenceFrameType>));
+    }
+    return sequenceFrames;
+  }
+
+  return new Image(view as unknown as BaseDataFields<ImageType>);
+};
