+++ ui/components/core/src/lib/utils/views/index.ts
@@ -1,0 +1,1 @@
+export * from "./createTypedView";
